requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.10
requests==2.32.3
drf-spectacular==0.28.0
